Add events to issue created/updated related events

+++ lib/eventum/class.history.php
@@ -75,7 +75,7 @@ class History
         DB_Helper::getInstance()->query($stmt, $params);
 
         $params['his_id'] = DB_Helper::get_last_insert_id();
-        $params['prj_id'] = Auth::getCurrentProject();
+        $params['prj_id'] = (int)Auth::getCurrentProject();
 
         $event = new GenericEvent(null, $params);
         EventManager::dispatch(Event\SystemEvents::HISTORY_ADD, $event);
